XML_Parse is well-defined as accepting only an int for the size arg
git-svn-id: https://svn.eu.apache.org/repos/asf/apr/apr-util/trunk@587432 13f79535-47bb-0310-9956-ffa450edef68

+++ xml/apr_xml.c
@@ -383,7 +383,7 @@ static apr_status_t do_parse(apr_xml_parser *parser,
         parser->error = APR_XML_ERROR_PARSE_DONE;
     }
     else {
-        int rv = XML_Parse(parser->xp, data, len, is_final);
+        int rv = XML_Parse(parser->xp, data, (int)len, is_final);
 
         if (rv == 0) {
             parser->error = APR_XML_ERROR_EXPAT;
